# doc/03_openapi_koipa_kl.yaml
openapi: 3.0.3
info:
  title: Koipa AI Engine API
  description: |
    한국지식재산보호원 AI 영업비밀관리시스템 — 한국지식재산보호원 AI 엔진의
    KL 웹시스템 연동용 REST API 명세.

    담당 기능: FUN-003, FUN-004, FUN-005, FUN-022, FUN-023, FUN-024
    호출 주체: KL 웹시스템 (Backend)
    제공 주체: Koipa AI Engine (FastAPI)

    인증: API Key (서버 간 통신) + 선택적 JWT (관리자 행위 추적)
    모든 응답은 application/json. 파일 업로드는 multipart/form-data.
    날짜·시각은 ISO 8601 UTC. 모든 ID는 UUIDv4.

    내부 RAG 백엔드: PostgreSQL pgvector dense + bigram tsvector `ts_rank`
    하이브리드. 본 명세는 backend-agnostic 하므로 KL 측은 RAG 동작 세부를
    몰라도 호출 가능. ES 경로는 레거시 어댑터로만 남아 있다.
  version: 1.0.0-draft
  contact:
    name: Koipa AI Engineering
servers:
  - url: http://koipa-api:8000/api/v1
    description: 온프레미스 Docker 내부망
  - url: https://{host}/api/v1
    description: 프로덕션
    variables:
      host:
        default: ai.example.koipa.local
security:
  - ApiKeyAuth: []
...
paths:
...
  /train:
    post:
      tags: [training]
      summary: 모델 학습·재학습 트리거
      description: |
        스케줄 또는 KL 관리자 수동 트리거. 진행 상태는 /train/jobs/{id}로 폴링.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [training_type, actor]
              properties:
                training_type:
                  type: string
                  enum: [full, incremental, hyperparam_search]
                base_model:
                  type: string
                  enum: [kf-deberta-base, koelectra-base]
                  default: kf-deberta-base
                dataset_version: { type: string, description: "MLflow dataset tag" }
                hyperparams:
                  type: object
                  additionalProperties: true
                use_rag: { type: boolean, default: false }
                actor: { $ref: '#/components/schemas/Actor' }
      responses:
        '202':
          description: 학습 작업 등록
          content:
            application/json:
              schema:
                type: object
                properties:
                  train_job_id: { type: string, format: uuid }
                  status_url: { type: string, format: uri }
                  websocket_url: { type: string, format: uri }
  /train/jobs/{train_job_id}:
    get:
      tags: [training]
      summary: 학습 작업 진행 상태 조회
      parameters:
        - in: path
          name: train_job_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: 학습 상태
          content:
            application/json:
              schema:
                type: object
                properties:
                  train_job_id: { type: string, format: uuid }
                  status:
                    type: string
                    enum: [queued, preprocessing, training, evaluating, registering, done, failed]
                  progress: { type: number, minimum: 0, maximum: 1, example: 0.42 }
                  current_epoch: { type: integer }
                  total_epochs: { type: integer }
                  metrics_so_far:
                    $ref: '#/components/schemas/Metrics'
                  started_at: { type: string, format: date-time }
                  estimated_finish_at: { type: string, format: date-time }
                  model_version: { type: string, nullable: true }
                  error: { type: string, nullable: true }
  /train/jobs:
    get:
      tags: [training]
      summary: 학습 작업 이력 목록
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 20, maximum: 100 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: 목록
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { type: integer }
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/TrainJobSummary' }
  /synth/generate:
    post:
      tags: [synthesis]
      summary: LLM 기반 학습용 샘플 문서 생성
      description: |
        FUN-003. 등급별·도메인별 합성 문서를 LLM으로 생성하고 검수 큐에 적재한다.
        Adapter로 OpenAI/Anthropic/Google/vLLM 교체 가능.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target_grade, count, actor]
              properties:
                target_grade: { $ref: '#/components/schemas/Grade' }
                domain:
                  type: string
                  enum: [tech, business, hr, finance, legal, mixed]
                  default: mixed
                count: { type: integer, minimum: 1, maximum: 500 }
                llm_provider:
                  type: string
                  enum: [anthropic, openai, google, vllm_qwen, vllm_exaone]
                  default: anthropic
                seed_documents:
                  type: array
                  items: { type: string, description: "Guide 문서 ID" }
                actor: { $ref: '#/components/schemas/Actor' }
      responses:
        '202':
          description: 생성 작업 등록
          content:
            application/json:
              schema:
                type: object
                properties:
                  synth_job_id: { type: string, format: uuid }
                  expected_count: { type: integer }
                  estimated_cost_usd: { type: number, format: float }
  /synth/queue:
    get:
      tags: [synthesis]
      summary: 검수 대기 중인 합성 문서 목록 (KL 검수 UI용)
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [pending, approved, rejected], default: pending }
        - in: query
          name: limit
          schema: { type: integer, default: 50, maximum: 200 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: 목록
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { type: integer }
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/SyntheticDoc' }
  /synth/{synth_id}/review:
    post:
      tags: [synthesis]
      summary: 합성 문서 검수 결과 등록 (승인/반려)
      parameters:
        - in: path
          name: synth_id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision, actor]
              properties:
                decision:
                  type: string
                  enum: [approve, reject]
                corrected_grade: { $ref: '#/components/schemas/Grade' }
                comment: { type: string, maxLength: 2000 }
                actor: { $ref: '#/components/schemas/Actor' }
      responses:
        '200':
          description: 검수 결과 반영
          content:
            application/json:
              schema:
                type: object
                properties:
                  synth_id: { type: string, format: uuid }
                  final_status: { type: string, enum: [approved, rejected] }
                  added_to_dataset_version: { type: string, nullable: true }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Grade:
      type: string
      description: 등급체계 코드. 관리자가 수정 가능하지만 기본 4단계 제공.
      enum:
        - TS    # 특급기밀
        - S1    # 1급 비밀
        - S2    # 2급 대외비
        - S3    # 3급 공개
...
    Actor:
      type: object
      required: [user_id, role]
      description: 호출자(관리자/시스템) 정보. 감사 로그용.
      properties:
        user_id: { type: string }
        role:
          type: string
          enum: [admin, reviewer, system, kl_backend]
        ip: { type: string, nullable: true }
...
    Metrics:
      type: object
      properties:
        accuracy: { type: number }
        precision_macro: { type: number }
        recall_macro: { type: number }
        f1_macro: { type: number }
        fnr_overall:
          type: number
          description: 미탐율. 보안 핵심 KPI.
        fnr_by_grade:
          type: object
          additionalProperties: { type: number }
...
    TrainJobSummary:
      type: object
      properties:
        train_job_id: { type: string, format: uuid }
        status: { type: string }
        training_type: { type: string }
        base_model: { type: string }
        started_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time, nullable: true }
        model_version: { type: string, nullable: true }
        f1_macro: { type: number, nullable: true }
        fnr_overall: { type: number, nullable: true }
    SyntheticDoc:
      type: object
      properties:
        synth_id: { type: string, format: uuid }
        target_grade: { $ref: '#/components/schemas/Grade' }
        domain: { type: string }
        llm_provider: { type: string }
        preview: { type: string, description: "본문 앞 200자" }
        generated_at: { type: string, format: date-time }
        status: { type: string, enum: [pending, approved, rejected] }
...